avoid leaking module augmentation for typescript
