+++ baselines/packages/wotan/api/src/runner.d.ts
@@ -1,6 +1,5 @@
 import { Linter } from './linter';
 import { LintResult, DirectoryService, MessageHandler, FileFilterFactory, Severity } from '@fimbul/ymir';
-import * as ts from 'typescript';
 import { ProcessorLoader } from './services/processor-loader';
 import { CachedFileSystem } from './services/cached-file-system';
 import { ConfigurationManager } from './services/configuration-manager';
@@ -19,11 +18,4 @@ export interface LintOptions {
 export declare class Runner {
     constructor(fs: CachedFileSystem, configManager: ConfigurationManager, linter: Linter, processorLoader: ProcessorLoader, directories: DirectoryService, logger: MessageHandler, filterFactory: FileFilterFactory, programStateFactory: ProgramStateFactory);
     lintCollection(options: LintOptions): LintResult;
-}
-declare module 'typescript' {
-    function matchFiles(path: string, extensions: ReadonlyArray<string>, excludes: ReadonlyArray<string> | undefined, includes: ReadonlyArray<string>, useCaseSensitiveFileNames: boolean, currentDirectory: string, depth: number | undefined, getFileSystemEntries: (path: string) => ts.FileSystemEntries, realpath: (path: string) => string): string[];
-    interface FileSystemEntries {
-        readonly files: ReadonlyArray<string>;
-        readonly directories: ReadonlyArray<string>;
-    }
 }
